Fix naturalearth update
The latest naturalearth admin-0 update removed the ISO_A2 value for Kosovo.
This resulted in wrong behavior of the country logic function.
To keep it compatible with the current version, the code re-injects the ISO_A2 value for Kosovo, even if it may be technically wrong.
A future update could remove the ISO_A2 properties and only rely on the country name.
But for now, this fix will be used.